+++ spec/acceptance/resource_default_orders_controller_spec.rb
@@ -3,7 +3,7 @@
 require "acceptance_helper"
 require "sidekiq/testing"
 
-resource "Resources::DefaultOrder" do
+resource "ResourceDefaultOrders" do
   include ActiveJob::TestHelper
 
   header "Accept", "application/vnd.api+json"
@@ -20,7 +20,7 @@ resource "Resources::DefaultOrder" do
     ResourceDefaultOrder.delete_all
   end
 
-  get "resources/default_order" do
+  get "resource_default_orders" do
     before do
       FactoryBot.create(:resource_default_order, resource: resource, language: language_en)
       FactoryBot.create(:resource_default_order, resource: other_resource, language: language_en)
@@ -79,7 +79,7 @@ resource "Resources::DefaultOrder" do
     end
   end
 
-  post "resources/default_order" do
+  post "resource_default_orders" do
     requires_authorization
 
     let(:valid_params) do
@@ -116,10 +116,12 @@ resource "Resources::DefaultOrder" do
     end
   end
 
-  delete "resources/default_order/:id" do
+  delete "resource_default_orders/:id" do
     requires_authorization
 
-    let!(:resource_default_order) { FactoryBot.create(:resource_default_order, resource: resource, language: language_en) }
+    let!(:resource_default_order) do
+      FactoryBot.create(:resource_default_order, resource: resource, language: language_en)
+    end
     let(:id) { resource_default_order.id }
 
     it "deletes the default order resource" do
@@ -132,7 +134,7 @@ resource "Resources::DefaultOrder" do
     context "when an incorrect ID is sent" do
       let(:id) { "unknownId" }
 
-      it "returns unprocessable entity" do
+      it "returns not found" do
         do_request
 
         expect(status).to be(404)
@@ -140,10 +142,12 @@ resource "Resources::DefaultOrder" do
     end
   end
 
-  patch "resources/default_order/:id" do
+  patch "resource_default_orders/:id" do
     requires_authorization
 
-    let!(:resource_default_order) { FactoryBot.create(:resource_default_order, resource: resource, language: language_en) }
+    let!(:resource_default_order) do
+      FactoryBot.create(:resource_default_order, resource: resource, language: language_en)
+    end
     let(:id) { resource_default_order.id }
     let(:valid_update_params) do
       {
@@ -180,7 +184,7 @@ resource "Resources::DefaultOrder" do
     context "when an incorrect ID is sent" do
       let(:id) { "unknownId" }
 
-      it "returns unprocessable entity" do
+      it "returns not found" do
         do_request(valid_update_params)
 
         expect(status).to be(404)
@@ -188,13 +192,15 @@ resource "Resources::DefaultOrder" do
     end
   end
 
-  patch "resources/default_order/mass_update" do
+  patch "resource_default_orders/mass_update" do
     requires_authorization
 
     let(:lang) { "en" }
     let(:resource_ids) { [] }
     let(:resource_type) { ResourceType.find(resource.resource_type_id) }
-    let(:params) { {data: {attributes: {lang: lang, resource_ids: resource_ids, resource_type: resource_type.name}}} }
+    let(:params) do
+      {data: {attributes: {lang: lang, resource_ids: resource_ids, resource_type: resource_type.name}}}
+    end
 
     context "with no lang param" do
       let(:lang) { nil }
@@ -245,7 +251,10 @@ resource "Resources::DefaultOrder" do
         end
 
         context "when sending more than 1 resource default order" do
-          let!(:resource2) { Resource.joins(:resource_type).where("resource_types.name != ? AND resources.id NOT IN (?)", resource.resource_type.name, resource.id).first }
+          let!(:resource2) do
+            Resource.joins(:resource_type).where("resource_types.name != ? AND resources.id NOT IN (?)",
+              resource.resource_type.name, resource.id).first
+          end
           let(:resource_ids) { [resource.id, resource2.id] }
 
           it "returns an array with more than 1 resource default order" do
@@ -263,8 +272,14 @@ resource "Resources::DefaultOrder" do
       end
 
       context "with previous resource default orders" do
-        let!(:resource2) { Resource.joins(:resource_type).where("resource_types.name = ? AND resources.id NOT IN (?)", resource.resource_type.name, resource.id).first }
-        let!(:resource3) { Resource.joins(:resource_type).where("resource_types.name = ? AND resources.id NOT IN (?)", resource.resource_type.name, [resource.id, resource2.id]).first }
+        let!(:resource2) do
+          Resource.joins(:resource_type).where("resource_types.name = ? AND resources.id NOT IN (?)",
+            resource.resource_type.name, resource.id).first
+        end
+        let!(:resource3) do
+          Resource.joins(:resource_type).where("resource_types.name = ? AND resources.id NOT IN (?)",
+            resource.resource_type.name, [resource.id, resource2.id]).first
+        end
         let!(:resource_default_order) do
           FactoryBot.create(:resource_default_order, resource: resource, language: language_en, position: 1)
         end
